+++ caliko-visualisation/src/main/java/au/edu/federation/caliko/visualisation/Line2D.java
@@ -130,7 +130,7 @@ public class Line2D
                                                                      GL_FLOAT,  // Data type
                                                                          true,  // Normalised? Colour values are between 0.0 and 1.0, so yes!
                                                 COMPONENT_COUNT * Float.BYTES,  // Stride
-                                             VERTEX_COMPONENTS * Float.BYTES);  // Offset
+                                        (long)VERTEX_COMPONENTS * Float.BYTES);  // Offset
 			
 			// Unbind VBO
 			glBindBuffer(GL_ARRAY_BUFFER, 0);
