+++ caliko-visualisation/src/main/java/au/edu/federation/caliko/visualisation/Point3D.java
@@ -148,7 +148,7 @@ public class Point3D
 	                                                                      GL_FLOAT,  // Data type
 	                                                                         false,  // Normalised?
 	                                                 COMPONENT_COUNT * Float.BYTES,  // Stride
-	                                              VERTEX_COMPONENTS * Float.BYTES);  // Offset
+	                                         (long)VERTEX_COMPONENTS * Float.BYTES);  // Offset
 				// Unbind VBO
 				glBindBuffer(GL_ARRAY_BUFFER, 0);
